requirements.txt
Adafruit-BBIO==1.0.1
Adafruit-GPIO==1.0.1
Adafruit-PureIO==0.2.1
appdirs==1.4.0
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
Werkzeug==0.11.15
argparse==1.2.1
click==6.7
distribute==0.7.3
packaging==16.8
pyparsing==2.1.10
pyserial==3.2.1
six==1.10.0
spidev==3.2
wsgiref==0.1.2
git+http://github.com/kevinjos/Adafruit_Python_BME280.git@master
